Rename disperserserver -> apiserver
