+++ docker-compose-build.yaml
@@ -8,7 +8,7 @@ services:
   disperser:
     build:
       context: .
-      dockerfile: disperser/cmd/disperserserver/Dockerfile
+      dockerfile: disperser/cmd/apiserver/Dockerfile
     image: ghcr.io/layr-labs/eigenda/disperser:${BUILD_TAG:-latest}
   encoder:
     build:
